# arch/arm/armv8-m : Fix performance issue due to cache setting