The syscall performance is poor since we are using no cache setting
in MPU. Change this to write back cache setting to improve performance.

Below is the performance comparison for both cases:

Syscall performance

Syscall type	| No cache case	| With WB cache case
syscall 0	| 52 secs	| 2 secs
syscall 1	| 59 secs	| 2 secs
syscall 2	| 51 secs	| 2 secs
syscall 3	| 184 secs	| 37 secs
syscall 4	| 57 secs	| 8 secs
syscall 5	| 66 secs	| 8 secs
syscall 6	| 282 secs	| 23 secs


Average context switch time

No cache : 138 usecs
WB cache : 19 usecs

Signed-off-by: Kishore S N <kishore.sn@samsung.com>