@sunghan-chang 2020-11-26
@kishore-sn Could you check the kernel testcase is working?