@jeongchanKim 2020-11-26
> @kishore-sn Could you check the kernel testcase is working?

I checked kernel testcase. It runs well with this PR.
```
########## Kernel TC End [PASS : 404, FAIL : 0] ##########
```