@kishore-sn 2020-11-26
@jeongchanKim 
Thanks for running the kernel tc. 
Generally, I would have run the kernel tc when any changes in the core kernel files. 
I will make it a point to mention the results in the PR from next time.